requirements.txt
flask
pytesseract
transformers
torch
textstat
scikit-learn
joblib
pdf2image
pillow
PyMuPDF
tensorflow
